Split components into the packages, and plugins into the extensions.
