+++ packages/codemirror/package.json
@@ -35,11 +35,15 @@
     "@jupyterlab/codeeditor": "^0.19.1",
     "@jupyterlab/coreutils": "^2.2.1",
     "@jupyterlab/observables": "^2.1.1",
+    "@jupyterlab/statusbar": "^0.7.1",
     "@phosphor/algorithm": "^1.1.2",
+    "@phosphor/commands": "^1.6.1",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/disposable": "^1.1.2",
     "@phosphor/signaling": "^1.2.2",
-    "codemirror": "~5.39.0"
+    "@phosphor/widgets": "^1.6.0",
+    "codemirror": "~5.39.0",
+    "react": "~16.4.2"
   },
   "devDependencies": {
     "@types/codemirror": "~0.0.46",
