+++ packages/codemirror/src/index.ts
@@ -13,6 +13,7 @@ export * from './mode';
 export * from './editor';
 export * from './factory';
 export * from './mimetype';
+export * from './syntaxstatus';
 
 /**
  * The default editor services.
