+++ packages/codemirror/tsconfig.json
@@ -17,6 +17,9 @@
     },
     {
       "path": "../observables"
+    },
+    {
+      "path": "../statusbar"
     }
   ]
 }
