+++ packages/docmanager-extension/src/index.ts
@@ -20,7 +20,9 @@ import {
   renameDialog,
   getOpenPath,
   DocumentManager,
-  IDocumentManager
+  IDocumentManager,
+  PathStatus,
+  SavingStatus
 } from '@jupyterlab/docmanager';
 
 import { DocumentRegistry } from '@jupyterlab/docregistry';
@@ -28,6 +30,8 @@ import { DocumentRegistry } from '@jupyterlab/docregistry';
 import { IMainMenu } from '@jupyterlab/mainmenu';
 
 import { Contents, Kernel } from '@jupyterlab/services';
+
+import { IStatusBar } from '@jupyterlab/statusbar';
 
 import { IDisposable } from '@phosphor/disposable';
 
@@ -84,7 +88,7 @@ const pluginId = '@jupyterlab/docmanager-extension:plugin';
 /**
  * The default document manager provider.
  */
-const plugin: JupyterLabPlugin<IDocumentManager> = {
+const docManagerPlugin: JupyterLabPlugin<IDocumentManager> = {
   id: pluginId,
   provides: IDocumentManager,
   requires: [ICommandPalette, IMainMenu, ISettingRegistry],
@@ -156,9 +160,74 @@ const plugin: JupyterLabPlugin<IDocumentManager> = {
 };
 
 /**
- * Export the plugin as default.
+ * A plugin for adding a saving status item to the status bar.
  */
-export default plugin;
+export const savingStatusPlugin: JupyterLabPlugin<void> = {
+  id: '@jupyterlab/docmanager-extension:saving-status',
+  autoStart: true,
+  requires: [IStatusBar, IDocumentManager],
+  activate: (
+    app: JupyterLab,
+    statusBar: IStatusBar,
+    docManager: IDocumentManager
+  ) => {
+    let item = new SavingStatus({ docManager });
+
+    // Keep the currently active widget synchronized.
+    item.model!.widget = app.shell.currentWidget;
+    app.shell.currentChanged.connect(
+      () => (item.model!.widget = app.shell.currentWidget)
+    );
+
+    statusBar.registerStatusItem('saving-status', item, {
+      align: 'middle',
+      isActive: () => {
+        return true;
+      },
+      stateChanged: item.model!.stateChanged
+    });
+  }
+};
+
+/**
+ * A plugin providing a file path widget to the status bar.
+ */
+export const pathStatusPlugin: JupyterLabPlugin<void> = {
+  id: '@jupyterlab/docmanager-extension:path-status',
+  autoStart: true,
+  requires: [IStatusBar, IDocumentManager],
+  activate: (
+    app: JupyterLab,
+    statusBar: IStatusBar,
+    docManager: IDocumentManager
+  ) => {
+    let item = new PathStatus({ docManager });
+
+    // Keep the file path widget up-to-date with the application active widget.
+    item.model!.widget = app.shell.currentWidget;
+    app.shell.currentChanged.connect(() => {
+      item.model!.widget = app.shell.currentWidget;
+    });
+
+    statusBar.registerStatusItem('path-status', item, {
+      align: 'right',
+      rank: 0,
+      isActive: () => {
+        return true;
+      }
+    });
+  }
+};
+
+/**
+ * Export the plugins as default.
+ */
+const plugins: JupyterLabPlugin<any>[] = [
+  docManagerPlugin,
+  pathStatusPlugin,
+  savingStatusPlugin
+];
+export default plugins;
 
 /* Widget to display the revert to checkpoint confirmation. */
 class RevertConfirmWidget extends Widget {
