+++ packages/docmanager-extension/tsconfig.json
@@ -26,6 +26,9 @@
     },
     {
       "path": "../services"
+    },
+    {
+      "path": "../statusbar"
     }
   ]
 }
