+++ packages/docmanager/src/index.ts
@@ -4,4 +4,6 @@
 export * from './dialogs';
 export * from './manager';
 export * from './savehandler';
+export * from './savingstatus';
+export * from './pathstatus';
 export * from './widgetmanager';
