+++ packages/fileeditor/package.json
@@ -34,9 +34,11 @@
     "@jupyterlab/apputils": "^0.19.1",
     "@jupyterlab/codeeditor": "^0.19.1",
     "@jupyterlab/docregistry": "^0.19.1",
+    "@jupyterlab/statusbar": "^0.7.1",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/messaging": "^1.2.2",
-    "@phosphor/widgets": "^1.6.0"
+    "@phosphor/widgets": "^1.6.0",
+    "react": "~16.4.2"
   },
   "devDependencies": {
     "rimraf": "~2.6.2",
