+++ packages/fileeditor/tsconfig.json
@@ -14,6 +14,9 @@
     },
     {
       "path": "../docregistry"
+    },
+    {
+      "path": "../statusbar"
     }
   ]
 }
