+++ packages/notebook-extension/src/index.ts
@@ -50,22 +50,22 @@ import {
   NotebookPanel,
   NotebookTracker,
   NotebookWidgetFactory,
-  StaticNotebook
+  StaticNotebook,
+  CommandEditStatus,
+  NotebookTrustStatus
 } from '@jupyterlab/notebook';
 
 import { IRenderMimeRegistry } from '@jupyterlab/rendermime';
 
 import { ServiceManager } from '@jupyterlab/services';
 
+import { IStatusBar } from '@jupyterlab/statusbar';
+
 import { ReadonlyJSONObject, JSONValue } from '@phosphor/coreutils';
 
 import { Message, MessageLoop } from '@phosphor/messaging';
 
 import { Menu } from '@phosphor/widgets';
-
-import { commandEditItem } from './modestatus';
-
-import { notebookTrustItem } from './truststatus';
 
 /**
  * The command IDs used by the notebook plugin.
@@ -283,6 +283,68 @@ const tools: JupyterLabPlugin<ICellTools> = {
   id: '@jupyterlab/notebook-extension:tools',
   autoStart: true,
   requires: [INotebookTracker, IEditorServices, IStateDB]
+};
+
+/**
+ * A plugin providing a CommandEdit status item.
+ */
+export const commandEditItem: JupyterLabPlugin<void> = {
+  id: '@jupyterlab/notebook-extension:mode-status',
+  autoStart: true,
+  requires: [IStatusBar, INotebookTracker],
+  activate: (
+    app: JupyterLab,
+    statusBar: IStatusBar,
+    tracker: INotebookTracker
+  ) => {
+    const item = new CommandEditStatus();
+
+    // Keep the status item up-to-date with the current notebook.
+    tracker.currentChanged.connect(() => {
+      const current = tracker.currentWidget;
+      item.model.notebook = current.content;
+    });
+
+    statusBar.registerStatusItem('command-edit-item', item, {
+      align: 'right',
+      rank: 4,
+      isActive: () =>
+        app.shell.currentWidget &&
+        tracker.currentWidget &&
+        app.shell.currentWidget === tracker.currentWidget
+    });
+  }
+};
+
+/**
+ * A plugin that adds a notebook trust status item to the status bar.
+ */
+export const notebookTrustItem: JupyterLabPlugin<void> = {
+  id: '@jupyterlab/notebook-extension:trust-status',
+  autoStart: true,
+  requires: [IStatusBar, INotebookTracker],
+  activate: (
+    app: JupyterLab,
+    statusBar: IStatusBar,
+    tracker: INotebookTracker
+  ) => {
+    const item = new NotebookTrustStatus();
+
+    // Keep the status item up-to-date with the current notebook.
+    tracker.currentChanged.connect(() => {
+      const current = tracker.currentWidget;
+      item.model.notebook = current.content;
+    });
+
+    statusBar.registerStatusItem('notebook-trust-item', item, {
+      align: 'right',
+      rank: 3,
+      isActive: () =>
+        app.shell.currentWidget &&
+        tracker.currentWidget &&
+        app.shell.currentWidget === tracker.currentWidget
+    });
+  }
 };
 
 /**
