+++ packages/notebook/package.json
@@ -39,6 +39,7 @@
     "@jupyterlab/observables": "^2.1.1",
     "@jupyterlab/rendermime": "^0.19.1",
     "@jupyterlab/services": "^3.2.1",
+    "@jupyterlab/statusbar": "^0.7.1",
     "@phosphor/algorithm": "^1.1.2",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/domutils": "^1.1.2",
