+++ packages/statusbar/context.ts
