+++ packages/statusbar/settings.ts
