+++ packages/statusbar/src/defaults/index.ts
@@ -4,7 +4,4 @@
 export * from './lineCol';
 export * from './kernelStatus';
 export * from './runningSessions';
-export * from './filePath';
-export * from './tabSpace';
 export * from './memoryUsage';
-export * from './savingStatus';
