+++ packages/statusbar/src/index.ts
@@ -6,3 +6,4 @@
 export * from './statusbar';
 export * from './style/statusbar';
 export * from './components';
+export * from './defaults';
