+++ packages/statusbar/src/style/lineForm.ts
